Fix flaky BackfillPartitionedWebHookLogsDaily spec
This MR takes inspiration from https://gitlab.com/gitlab-org/gitlab/-/merge_requests/182397
to attempt to fix the spec flakiness issue described in
https://gitlab.com/gitlab-org/gitlab/-/issues/520378.

The spec tries to save two web hook logs: one for last month, one for
yesterday. Specs are failing when trying to create yesterday's log,
with an error stating that no appropriate partition exists for that
date.

This MR ensures a partition exists for the current month.