Fix lookup of find_diff_start_pipeline_with_security_reports
When no common ancestor pipeline is found for scan finding policy
evaluation, we look for pipelines matching `diff_start_sha`.

This method uses `has_sbom_reports?` to lookup the pipelines.
We should use `has_security_reports?`, as the method name suggests.

This leads to cases when the target pipeline is not found.

Changelog: fixed
EE: true