No-op ci_runner_machines_687967fa8a table backfill migration
The backfill on this table is causing issues due to a problem in the
timing of the post-deployment migrations:

1. 20241107064635_queue_backfill_ci_runner_machines_partitioned_table
2. 20241211072300_retry_add_fk_from_p_ci_runner_managers_to_p_ci_runners
3. 20241219100359_requeue_backfill_ci_runners_partitioned_table

Migration 2 installs a FK check before the target table
(ci_runners_e59bb2812d) is backfilled in migration 3,
so we have https://gitlab.com/gitlab-org/gitlab/-/issues/520092.
We can afford to not backfill ci_runner_machines_687967fa8a,
since it'll be populated with runner data automatically as they
request jobs from the GitLab instance.

Changelog: other