graphql: Fix RepositoryBlob cache key
As far as the frontend is concerned, the RepositoryBlob type essentially
represents a file within a Git repository. This type supplies the data
used to render the file content and associated metadata in the GitLab
repository UI (the raw content itself, filename, file mode etc.).

The type is populated through queries such as getBlobInfo [1], which is
executed by the Vue BlobContentViewer component when rendering a file in
the GitLab UI. The results of the query are cached by the Apollo Client
[2], so queries for the same blob are fetched from the browser's cache.

By default, the cache key is computed by hashing the __typename and id
fields [3] of the type. In the case of `RepositoryBlob`, this would be
"RepositoryBlob" and something like
"gid://gitlab/Blob/ccbacb5ab000bbe86b541ba0d589dfcd62440e96". The object
field is then replaced [4] with a reference to the cache key. This is
problematic because (__typename, id) doesn't uniquely identify a blob
in the GitLab UI. It _does_ represent a blob in Git, because Git is
content-addressable and deduplicates objects based on content, but it's
legal to have many files in a repository with the same content but
_different_ names. The default fields used to compute the key doesn't
take this into account.

This wouldn't be a problem if the RepositoryBlob type only contained
the raw blob data and metadata common to all representations of the
blob. Unfortunately, the object also contains a series of filenames and
paths that are unique to a particular file representing that blob. This
means that hash collisions will occur, and the wrong data will be
returned from the cache. A typical flow looks like this:

0. Make sure the repository contains two files with the exact same
   data. For example purposes, we have file-a and file-b, both with the
   same blob ID of `deadbeef`. Let's say that Apollo computes the cache
   key to be `baddcafe`.
1. Browse to file-a via the GitLab UI. Verify that the filename
   displayed above the content matches with the breadcrumbs at the
   top of the page. This action causes a getBlobInfo query that will
   cache the `RepositoryBlob` using `baddcafe` as the key. The cached
   object will contain `file-a` as its filename.
2. Click on the Back button in your browser, and navigate to file-b.
   This action causes another getBlobInfo query that overwrites the
   cache using the same `baddcafe` key. This time, the cached result
   will contain `file-b` as its filename.
3. Navigate back to file-a. This action does not cause a getBlobInfo
   query as the cache is still warm. Instead, the cached object with the
   key `baddcafe` is returned, and this object will be the incorrect one
   that was cached in step 2. This causes the UI to render the filename
   as `file-b` even though you're looking at `file-a`.

Prevent this cache collision from occurring by overriding the default
keys used by Apollo to compute the cache key. Instead of
(__typename, id), we'd ideally like to use a composite key of
(id, path), however this is reliant on the specific query including the
`path` as a response field. This is not always true, and results in
errors like

    Missing field 'path' while extracting keyFields from <object>

To overcome this, we can supply a function [5] to keyFields which
checks if `path` is available, and only incorporates it into the key if
so. This resolves the original issue when browsing files in the
repository tree which remaining compatible with queries which don't ask
for the `path`.

[1] https://gitlab.com/gitlab-org/gitlab/blob/bab972ade8385bc7e10cfdc86fce21ac97c7a046/app/graphql/queries/repository/blob_info.query.graphql#L1-1
[2] https://www.apollographql.com/docs/react/caching/overview
[3] https://www.apollographql.com/docs/react/caching/overview#2-generate-cache-ids
[4] https://www.apollographql.com/docs/react/caching/overview#3-replace-object-fields-with-references
[5] https://www.apollographql.com/docs/react/caching/cache-configuration#typepolicy-fields