Remove some options and indirection
Always use document as scrolling, since it doesn't work in modal
or sidebar.

Remove funciton to create resize observer. Ideally we shouldn't
have a multiple of these, but this is not a big deal at this time
so instead make the code easier to read and get rid of extra
event thing there.