Implement AccessibleOwnersFinder logic
This change implements logic to find accessible and inaccessible code
owners.

It also implements a base_finder which will be reused for the rest of
the finders going forward.

We load all of the users and groups that match owner names and emails.

Any names and emails that do not have a matching group or user are
marked as invalid.

Accessible groups' full_paths are assigned to the finder's
valid_group_names attr.

Accessible users' usernames are assigned to the finder's valid_usernames
attr.

In a follow-up MR this class will be executed in the CodeOwners::Process
and then used to assign errors to the CodeOwners::File instance.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261