Json Import Service for Compliance Frameworks
This commit adds a json import service for compliance
frameworks. This allows users to create a compliance
framework from a json template. It modifies the create
services for frameworks, requirements, and controls to
support the json import.

This work is for issue: https://gitlab.com/gitlab-org/gitlab/-/issues/513638

Changelog: added
EE: true