gitaly: Update supported Postgres versions
The Praefect documentation still lists Postgres 11 as the minimum
supported version. This was true when 11 was also the minimum supported
version across the broader GitLab product [1], but this is no longer the
case.

To avoid confusion and to provide consistent documentation, remove
explicit mentions of the supported Postgres versions. Instead, link to
the main Postgres documentation for GitLab.

[1] https://gitlab.com/gitlab-org/gitlab/-/commit/3231c944752e8339c41eb40e18ee2f4320f944cd