Remove usage of expose_url for attachment_url
In the Test Case object, stop using the expose_url function on the
attachment_url attribute.  The inclusion of the call would lead to
duplication of the relative_url_root in the resulting path.

Add documentation based on the review suggestion that expose_url
should only be used in API related calls, not Rails ones.

Use the file_project_job_artifact_url function to find the URL
instead of _path