Make bulk data seeder spec not_owned
The DataSeeder (bulk_data) was added as part of a working group
https://handbook.gitlab.com/handbook/company/working-groups/demo-test-data/.

It seeds test data into instances for tests or demos, but it does not
relate to the scalability of the GitLab application. Because it is
used by engineers that want to set up an environment, I think the ideal
group to own this would be the [Engineering Productivity
group][0]. But at the time of writing, this group does not contribute
to any feature categories, so applying not_owned for now.

In the context of https://gitlab.com/gitlab-org/gitlab/-/issues/499227

[0]: https://gitlab.com/gitlab-com/www-gitlab-com/-/blob/41452e0e04d0676a6337180c5533feb4a25235ad/data/stages.yml?page=3#L2998-3010