Rename code owner finders to filters
It was also suggested that we rename `Finder` as this is already a
recognized abstraction within the gitlab application. We use finders to
generate database queries.

While these objects do query database records, they do more than this.
They filter the input data into outputs. As this is performing a
filtering role we've landed on `Filter` as the naming preference.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/526519+

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261+
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/521151+