More comfortable styles for custom wiki sidebar content
Currently, users can create a custom sidebar markdown file for wikis,
but the rendered markdown is not styled to take into account the fact
that this content is displayed in a narrow aside.

This MR applies specific styling to custom sidebar content so that it is
more ergonomic.

- Reduce sizes of headings
- Reduce left-padding of lists

Changelog: changed