Fix terraform widget report sorting (again)
Prior attempts (by me) completely failed to notice how the existing
tests and test data used integer numbers, and my local testing didn't
have enough data to tease out these issues. Now, I'm using the existing
test data with new records, and the existing tests to actually check and
enforce the changes, and made sure the test failed before and passes
after with a little more carefulness.

The problem with the previous attempts is the same - cases where the
report might seem invalid but shouldn't be, and vice-versa, just due to
bad comparisons against the expected values of create/update/delete in
the returned report from the server. This commit re-uses the logic from
the original (before my changes) code, and just merges the check for
tf_report_error in a bit earlier. I've also undo my change to replace
the parsed report values with parsed values, as that shouldn't be needed
and wasn't helpful. Overall, this makes the code a lot smaller and is
the change I wish I'd originally done.