Update "Using Docker images" CI docs section
Existing documentation for how to use images from private container
registries lists requirements for those, who would like to use
`DOCKER_AUTH_CONFIGS` approach with either static credentials,
credentials store or credentials helpers. The requirements list mentions
that this feature is available for Kubernetes executor starting from
GitLab Runner version 13.1.

We've added this mention as before this feature was available only for
the Docker executor. And in 13.1 we've made it available for both
executors.

Current wording, however, may suggests that this feature is available
for only the Kubernetes executor, while it's not true. It's available
for Docker executor as well, and in fact - it predates its by few years.

This commit adds mention about the Docker executor to remove such
possible confusion.

Version v12.0 was chosen as this was the first version of GitLab Runner
that supported all ways of using `DOCKER_AUTH_CONFIGS`. Initial support
for it was added two years earlier. But given that it's currently an 8
year old version of GitLab Runner, and v12.0 is 6 year old version of
GitLab Runner - so still very old - I think we can just point to v12.0
for simplicity of docs and maturity of the feature.

This commit adds also missing history entries to the `Use
statically-defined credentials`, `Use a Credentials Store` and `Use
Cedential Helpers` sections.