Support SHA1 or SHA256 fingerprint for Group SAML
In the upcoming ruby-saml 2.x upgrade, the default IdP cert fingerprint
algorithm will change from SHA1 to SHA256. Groups use only
SHA1 fingerprints and will need to explicitly set SHA1 as the
algorithm after the upgrade. This change ensures we detect
the algorithm and set it for installs for backward compatibility.