Document IMDS V2 hop limit and other troubleshooting information
We had one customer running GitLab inside a Docker containern see
delays in retrieving instance credentials due to the hop limit of the
IMDS v2 request. This commit documents the need to set the IMDS hop
limit when running GitLab in a container. In addition, the
`disable_imds_v2` option is available in fog-aws to disable IMDS v2.

This commit also improves the Rails Console commands for debugging
problems.

Relates to
https://gitlab.com/gitlab-com/request-for-help/-/issues/2526

Changelog: changed