Expire session from time of creation
Adds a new application setting, hidden behind a feature flag. If
enabled, sessions expire after the "default session duration" time
specified in application settings. The sessions expire even if a user is
actively browsing GitLab, causing the user to have to sign in again. For
some organizations running GitLab, this is a necessary security control.

Co-authored-by: John Parent <john.parent@kitware.com>