Does not use specific prompt version on Duo Self-hosted
Developers can specify a version for their prompts in order to test
changes to the base prompt with feature flags. However, if the feature
flag becomes default on, then self-hosted models might fail as that
specific version might not exist. To overcome this, self-hosted will not
use fixed versions, only default pinned ones.