Add LFS troubleshooting with .gitattributes
We've have had several customers encounter an issue where LFS files
were no longer checked out after upgrading to GitLab Runner v17.7.0
and higher.

This is due to a `git-lfs` update in v3.6.0:
https://github.com/git-lfs/git-lfs/pull/5699.

LFS pointers that were not listed in `.gitattributes` could still be
identified as such in `git-lfs` v3.3.0, but this is no longer the case
in v3.6.0. Users have to check that all files are explicitly matched
in `.gitattributes`.

For example, suppose you had a repository with these files:

```
test1.bin
TEST2.BIN
```

Let's say `.gitattributes` were set to:

```
*.bin filter=lfs diff=lfs merge=lfs -text
```

Let's also assume that the repository were cloned on a case-sensitive
filesystem, like ext4. As a result, the Git config `core.ignorecase`
is `false`.

In `git-lfs` v3.3.0, `git-lfs` runs `git ls-tree -r <SHA>` to find
all files in the repository. In the example above, this includes
both `test1.bin` and `TEST2.BIN`, so these filenames are considered
as possible candidates for LFS pointers.

With this change, in `git-lfs` v3.6.0, `git-lfs` now runs something
along the lines of `git ls-files -r ':(top,attr:filter=lfs)'`, which
filters out files that only are matched by `.gitattributes`. Since
`TEST2.BIN` does not match the `*.bin` rule, it is ignored.

The solution, of course, is to update the `.gitattributes` [as
described in this issue](https://github.com/git-lfs/git-lfs/issues/4110#issue-607559373):

```
*.[bB][iI][nN] filter=lfs diff=lfs merge=lfs -text
```

Of course if `core.ignorecase` is `true`, then the previous
`.gitattributes` matches, and there are no issues.