vue: Use revised lastCommit resolver
de93de3f4c1a (graphql: Don't get the commit from paginatedTree, 2025-03-20)
modified the GraphQL resolver responsible for the lastCommit object.
This is the follow-up commit which will begin using the improved
resolver.

Modify the pathLastCommit query such that we no longer ask for a tree.
The `lastCommit` block is hoisted up as a child of `repository`, and the
`paginatedTree` block is removed completely.

# Conflicts:
#	app/graphql/queries/repository/path_last_commit.query.graphql

# Conflicts:
#	app/graphql/queries/repository/path_last_commit.query.graphql
#	spec/frontend/repository/components/last_commit_spec.js