Delete patch branch when Vulnerability Resolution creation fails
While trying to root cause the error occurring in the linked issue,
we realized that we are creating and abandoning branches regardless of
the root cause.

This is a poor and confusing customer experience.

This change deletes the generated code when the patch fails.

The fix for the root causes identified in the linked issue require
much more substantial changes, and will need to be fixed in separate
commits.

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/497193
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185761
Changelog: fixed
EE: true