Fix flaky test -- second try
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185272+, we
incorrectly fixed a flaky test thinking that the one date operated in
UTC and the other in localtime.

The issue, however, is different. It seems like both `freeze_time` and
`time_travel_to` doesn't resolve the problem. `create_token` operates in
a different context. Because of that, we decided to compare with `<=`
operator instead of `eq`.