Improve CodeOwner section parsing logic
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/525881+

Prior to this change:
- `required approvals` could not contain spaces
- `default owners` required a leading `[Doc] @owner` or `[Doc][2] @owner`

Now we parse these successfully but add an error.

This also improves the default owners regex by removing a duplicated
range i.e. `\w` already includes `_`

The looks like a section but it's not parsable regex is now simpler, we
just check if the line begins with `[` or `^[`.

This also restructures the class a bit so parsing and validation happens
during initialization. This allows us to expose some convenience methods
to make the consumption of this parser more readable

1. `section_header?` -> `true` if the line was successfully parsed
2. `unparsable_section_header?` -> `true` if `section_header?` is
   `false` but the line starts with `[` or `^[`

This cleans up the consumption side in `CodeOwners::File` too. Notably,
we skip attempting to parse the line as an entry if
`unparsable_section_header?` returned true.

Changelog: changed
EE: true