Prevent unwanted todo count update in the sidebar
In a previous change, we tried to consolidate the todo count updates by
tying the todo app's count update query to the global user counts store.
This ended up introducing a small bug where filtering the todos in the
todo app causes the sidebar's count to display the filtered count when
it should always show the total pending count.