Refactor code owner validation filters
We have some differences in the filters e.g. supplied input users or
groups vs none supplied, handling groups and users in the same filter
vs only users or groups.

Originally we implemented a base class, however, as we have these
variances we have decided to remove the base class.

We have also split the accessors into their own methods instead of
having an execute method. This allows the filters to behave more like
true objects.

We have also removed the need to define a constant for the error
message. As symbols are immutable in ruby there's not much benefit of
adding constants here. The old implementation was fairly abstract so
harder to grok. This new implementation makes it very clear what is
happening without needing to flick between the base and derived classes.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/526521+

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261+
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/521151+