Revert coverage image to previous
In the slim image, the `rugged` gem fails to build its native extensions
(https://gitlab.com/gitlab-community/gitlab/-/jobs/9558171729#L83).

Thats because the slim image doesn't contain CMake which is required
for this. It didn't fail in the canonical pipeline because of the gem cache
which is currently broken in the community fork. This problem will be
fixed separately as part of https://gitlab.com/gitlab-org/gitlab/-/issues/529053