Delete ci build trace metadata record after archive
When a CI build trace log is archived, a
trace metadata record is generated to keep
track of archival attempts. Once it succeeds,
however, we don't need this record anymore.
So this MR adds a process to the delete that
record on successful trace archival.

Behind a feature flag that's disabled by
default: ci_delete_archived_trace_metadata