Fix bug when secrets provider URL given in dotenv
Originally, we had a bug where the JWT tokens for the build had
incorrect value for expiry - "exp". It did not respect the values of
runner timeout and job timeout. This was because the JWT tokens were
computed and memoized at a stage where it has incomplete information
about the timeout. This happened as a side effect of the way we were
validating the presence of secrets provider URL.
The details of this bug can be found here -
https://gitlab.com/gitlab-org/gitlab/-/issues/351765
This bug was fixed in this MR
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/130769
The fix involved moving the validation of the secrets provider URL to
the pipeline creation phase. However this fix introduced a regression
where the pipeline creation would fail if the secrets provider URL was
provided as a dotenv variable from the output of a different job.
This was because at the time of pipeline creation, we DO NOT have access
to dotenv variables generated only at job run time.

In this commit, we revert the bug fix MR that caused the regression and
put back the validation for the secrets provider URL to the
job creation phase. And to solve the original bug, we DO NOT compute and
memoize the JWT tokens when we DO NOT have timeout information. We do
this by checking if the secrets provider URL is present ONLY in the
subset of variables where it can be present. Only these subset of
variables are computed and memoized at the time of Secrets Provider
validation. All 3 changes are behind Feature flags

References:
[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/469137)

Changelog: fixed
EE: true