Protected packages: Add minimumAccessLevelForDelete in proj. settings UI
This MR integrates the attribute `minimumAccessLevelForDelete`
to the project settings UI. This attribute is used
to set the minimum access level required to delete a package.
With this, project owners can use
the frontend to set the minimum access level.

The attribute is similarly integrated as
the other attribute `minimumAccessLevelForPush`.

Changelog: added