Return the namespace cluster agent mapping after it is deleted
Previously, when delete mutation was called on a namespace cluster
agent mapping, only errors field would be returned in the response.
The service layer also did not have return any information related
to the deleted mapping.

This change returns the deleted mapping in the service layer as well
as the GraphQL layer.

Changelog: changed
EE: true