Update WORK_ITEM_TYPE_VALUE_* to WORK_ITEM_TYPE_NAME_*
The strings for these constants come from `TYPE_NAMES` in the backend
in `app/models/work_items/type.rb`, and they are exposed in GraphQL
in the field `name`. Therefore, these constants are being renamed
to `*NAME*` for consistency so when working with the code we don't
need to keep translating `*VALUE*` to the `name` strings in our heads.