Implement EligibleApproversFinder logic
This change implements logic to find eligible approvers for code owner
entries. It is the first step in refactoring the
`Gitlab::CodeOwners::UserPermissionCheck` logic.

We select all of the input_users which can update_merge_request which is
the ability required to approve an MR.

This change also improves on the `UserPermissionCheck` by including
checks for the supplied emails.

To ensure this check is performant we preload everything required to
perform these checks.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261+