Apply inaccessible owner errors to code owners file
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/184798+ we
introduced a filter to locate all of the accessible and inaccessible
code owners.

In this change we are executing this filter and then applying errors to
the code owner file for each of the inaccessbile owners.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261+