Set cache: [] in Secret Detection latest template
When using global cache in GitLab CI, SD may scan cached dependencies
and find false positives from outside of the Git repository. This change
disables cache in the latest secret detection template to prevent this
and improve performance.

This change only affects the latest template and will be introduced to
the stable template in 19.0, as it is a breaking change.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186637