Stores instance ai logging config to database
Instance admins were configure AI logging using the UI that would enable
or disable a feature flag. This was allowed us to add the UI quickly,
but at the cost of technical debt, since the Feature Flag is intended to
be used to manage logging for projects on the .com instance. Moving to
an instance setting decouples the behavior.

Changelog: changed