Return empty array when no PATs for service account
All other access tokens endpoints return `[]` when the query returns not
tokens.

Here we change the behavior of the group service account listing the
personal access tokens (PATs).

From: `{ "message": "404 Personal access token(s) Not Found" }`

To: `[]`

In addition, we fix the frontend to replace the `:id` placeholder for
the service account ID.