Add support for deleting agent managed resources on environment stop
When using GitLab-managed Kubernetes resources with ephemeral environments
(for example, review apps), many resources can be created across many namespaces.
When the environment is no longer required and stopped/removed in GitLab, the
Kubernetes objects are left in the cluster, potentially consuming resources.

This change supports a new environment template configuration option,
`delete_resources`, which is used to indicate that resources should be removed
when the environment is stopped.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/188024

Changelog: added