Introduce read_group_metadata ability
read_group_metadata ability gives a user access to minimal information
about a group (currently id, name, full_name, web_url, avatar_url,
user_permissions). It is enabled when read_group is enabled.

In EE, this ability is granted to a user that is not a member of a group
but has custom admin role with read_admin_cicd ability enabled.