Prevent unwanted todos refetch after snoozing
This ensures that the snooze todo mutation's optimistic response matches
the server's response. Until now, there was a format mismatch in the
`snoozedUntil` field's format where the optimistic reponse included the
milliseconds while the server had them stripped. We now strip the
milliseconds in the client-side to ensure cache consistency, which in
turns prevents the `todos` query from being triggered unexpectedly.