Rename main plan to base product in namespace provision API endpoint
Prior to this change, the namespace provision API endpoint was using a
param called `main_plan` to sync subscription plan related info.
However, the term is outdated and using `base product` is actually
preferred.

This change updates the `main_plan` param to `base_product` in the
namespace provision API endpoint. The endpoint is not actively used at
the moment as development for it is still ongoing and any implemented
features are still behind a disabled feature flag.

Changelog: changed
EE: true