Filter out merge requests in lists with Duo bot reviewers
For the returned to you list we will only return merge requests where there are reviews
left and that the only review left is not from the bot user.
For the assigned to you list we will return merge requests that have no reviewer
or if the only reviewer is the Duo bot user.

https://gitlab.com/gitlab-org/gitlab/-/issues/521381