Apply owners_without_permission errors to CODEOWNERS
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185299+ we
reimplemented the UserPermissionsCheck into the EligibleApproversFinder.

In this change we are executing this filter and then applying errors to
the code owner file for each of the owners which do not have permission
to approve a merge request.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261+