Add Backend Support for finding_token_status
Adds finding_token_status as a field that gets passed
as part of vulnerability_finding_data and can be used
on the Frontend in a future MR to display the token
status. Updated finding factory to allow creation
of finding_token_status for rspec tests and added
test.

Changelog: added
EE: true