UI for JSON Import Service
This adds a UI for the json import service for compliance
frameworks. This allows users to create a compliance
framework from a json template. If there is an error
with the import, the service does a best effort of
creating the framework and reporting back specific errors.

This work is for issue: https://gitlab.com/gitlab-org/gitlab/-/issues/513640

Changelog: added
EE: true