Introduce read_project_metadata ability
read_project_metadata ability gives a user access to minimal information
about a project (currently id, name, description, web_url, avatar_url,
user_permissions). It is enabled when read_project is enabled.

In EE, this ability is granted to a user that is not a member of a
project (or any of its ancestor groups) but has custom admin role with
read_admin_cicd ability enabled.