Prevent invalid tokens from falling back to Git password auth
When authenticating with Git over HTTPS, authentication falls
through a number of mechanisms since Git password may be a token,
local database password, LDAP password and more. This change
ensures that any recognized token type does not fallback to
database or LDAP authentication, for example, when the token
is revoked or expired.

Changelog: fixed