Make it clearer that index must come before foreign key
I noticed these docs say that they can be in the same migration but this
would actually be bad if the foreign key is added first in the same
migration which uses `disable_ddl_transaction!`. That's because the
foreign key will exist before the index and we have all the same
problems documented here.