Handle errors when smartcard user not found on LDAP server
Currently if you try to log in using a smart card for an LDAP server,
but the certificate doesn't return anything from the LDAP server, an
error is raised. With this change, the user will be redirected to the
sign-in page with the message "failed to log in," the same as if the
user entered an incorrect name or password for LDAP sign-in.

Changelog: fixed
EE: true