Add license approvals for prawn pdf gems
Context
-----------------

We want to use `prawn`, `prawn-table`, and `prawn-svg` to build out
PDF reporting features for our security features.

We previously built a proof of concept, and have decided to move
forward implementing the report with `prawn`[^1]

Before beginning the development work, we need to:[^2]

- [x] get legal approval for the gem licenses [^3]
- [ ] adds the license approvals to the license scanner (this change)
- [ ] add the approvals to omnibus
    - maybe? I am a little unclear if this is required for pure ruby dependencies
    - https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8249+

This MR
-----------------

This adds the license approvals. It is broken out from the [MR][4]
where the gems are being added. The docs are a bit vague on what order
everything needs to be done. It seems like having the approvals in
`master` first will simplify the rest of the process.

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/524059
Changelog: added

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/183327
[1]:https://gitlab.com/gitlab-com/legal-and-compliance/-/issues/2565
[2]:https://handbook.gitlab.com/handbook/legal/product/#using-open-source-software
[3]:https://docs.gitlab.com/development/licensing/#getting-an-unknown-or-lead-licensed-software-approved
[4]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/184931

[^1]: The proof of concept [MR][0] builds out a demo using the gems
[^2]: These steps come from the ["Getting an Unknown Licensed Software Approved"][3] and Legal's handbook page on
      ["Using Open Souce Software"][2]
[^3]: License approval was already obtained [here][1]