Add Duo Nano support to provision API and service for GitLab.com
We extend the internal API to allow CustomersDot to provision Duo Nano
add-on purchases at a GitLab instance. Duo Nano gets provisioned in the
same way other Duo add-ons get provisioned. The structure for the API is
also the same we use for other Duo add-ons, for example with attributes
like started_on and expires_on.

Changelog: added
EE: true