Protected containers: Minimum access level delete in project setting UI
This MR integrates the attribute `minimumAccessLevelForDelete`
for the container respository protection rules
in the project settings UI.
This attribute is used to set the minimum access level required to
delete a container repository.
With this, project owners can use the ~frontend to
set the minimum access level.

The attribute `minimumAccessLevelForDelete` is similarly integrated as
the other attribute `minimumAccessLevelForPush`.

The implementation in this MR also integrates / considers
the feature flag `container_registry_protected_containers` by:
- Hiding the field "Minimum Access Level For Delete" in the form for
entering a new container repository protection rule
- Hiding the default option (`Developer (default)`)
for the field "Minimum Access Level For Push" in the form
for entering a new container repository protection rule
- Hiding the column "Minimum Access Level For Delete"
in the table listing the container repository protection rules
- Hiding the default option (`Developer (default)`)
for the field "Minimum Access Level For Push"
in the table listing the container repository protection rules

Changelog: added