Apply unqualified_group errors to CODEOWNERS
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186117+ we
implemented logic to filter groups that were eligible to be code owners.
This includes the project's group and its ancestors, and any group
shared with the project with at least developer access.

In this change we are executing this filter and then applying errors to
the code owner file for each of the group owners which are not eligible.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/521151+