Add group_without_eligible_approvers code owner validation error
Adds new error to blob/constants.js to handle new code owners validation
error added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186406+

Adds new mock data which is used in the codeowners validation spec to
ensure the errors are rendered.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/521152+

The logic to implement the error will be added in a follow up. This
error needs to be added to the frontend before we implement the change
or the frontend will render `undefined`.