Clarify access to PEP configuration when using $CI_JOB_TOKEN
When using $CI_JOB_TOKEN to create downstream pipelines, the setting
"Pipeline execution policies" to grant access automatically is
not enough to grant access. The job token scope has to be additionally
configured for the security policy project
to allow the groups or projects to access it and
read the pipeline execution policy CI configuration.

Changelog: changed