Return a success status when stopping an already stopped environment
This is a follow-up to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/181339,
which returned a bad request in the environments API when an environment fails to
stop. This change unintentionally caused error responses when attempting to stop an
environment that was already stopped or stopping, which is not the expected behaviour.

Changelog: changed