Backfill AdminRole validations and callbacks
In an effort to bring parity to AdminRole, I've extracted a concern here
and am doing some duck typing to share code between AdminRole and
MemberRole. Part of the reason for this is to catch any code-churn that
could occur when MemberRole moves forward while the AdminRole refactor
is still underway.

I believe that if we can get AdminRole to quack like MemberRole that
this entire refactor will go much more smooth.

* extract Authz::MemberRoleLifecycle from MemberRole and share with
  AdminRole

part of:
https://gitlab.com/groups/gitlab-org/-/epics/17390