Make trie docs use explicit example
When I read these docs the first time I found them a little unclear
because the input to Trie was not explicit but all the other arguments
were using explicit examples. I was struggling to work out exactly what
input is needed to get that output. It was in the comments but they
weren't totally clear to me either.