Revert support for multiple keys in attr_encrypted
Support for multiple encryption keys was recently added, but in practice
we still don't allow to define multiple keys in `config/secrets.yml`.

Since there's an MR to use the upstream attr_encrypted gem again:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186797
and that we can use Active Record Encryption (which support multiple
encryption keys), we're reverting the multiple keys support in
attr_encrypted here, so that we can use the upstream gem.

We should migrate all attr_encrypted fields to Active Record Encryption
in the near future.

Signed-off-by: Rémy Coutable <remy@rymai.me>