Update table_size database dictionary entries
Updates database dictionary entries for `vulnerability_identifiers`

You can read more about our process to classify table size in
https://docs.gitlab.com/ee/development/database/large_tables_limitations.html.

Verify this MR as it was automatically created by `gitlab-housekeeper`.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::UpdateTableSizes keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: added