Add a new ResourceGroup process_mode: "newest_ready_first"
The problem with newest_first in situations with frequent merges and
continuous deployment is that you can end up in a situation where the
"newest" pipeline job is forever waiting on the next-added, and you get
zero deploys.

This new mode will still grab the newest job, but it will not wait for
created or scheduled jobs, instead only grabbing jobs which are in
waiting_for_resource state. This means that even in the face of lots of
new jobs that are (for ie) in the state "created" you will still get
deploys, but when multiple pipelines reach the waiting for resource
state while another job has claimed the resource, only the newest one
will actually run.

Changelog: changed